Fix logic that is used when reading restart files.
Without this change, CESM exact restart tests failed when run with
daily coupling, as is the case for compsets other than TG. For
example, this test was failing:

ERS_Ly5.T31_g37.IGCLM45.yellowstone_intel

The problem was that glimmer_nc_checkread was first called for the glide
variables. In that call, sub_time was set to model%numerics%time at the
BEGINNING of the routine (0), and NCI%processed_time is set based on that. But
then model%numerics%time is updated in that routine. Then glimmer_nc_checkread
is called for the glint variables. Now sub_time is the updated value, so this
comparison:

    if (sub_time > NCI%processsed_time) then

is true when it should be false.

The fix, implemented here, is to turn sub_time into a function, which is called
whenever needed. That way, NCI%processed_time is set to the updated version of
sub_time.

---

Testing done:

- Ran full CESM aux_glc test suite

- Tested build of standalone code on hopper, using the master_build_hopper script

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@997 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 69fa960134c7a9d093ba2550f0b895888c74baf5