MAKE CLEAN REQUIRED!
This commit undoes some of the changes I made in r761.  Specifically, in that commit I made all unused variables to be allocated to size 0.  Here I have reverted that to simply leave them unallocated.  This simplifies the logic for allocating variables and should be easier to maintain.

Specifically, I have changed the is_enabled_*() functions in ncdf_template.F90.in to use the associated() function instead of the size() function.  Then, I also removed the allocation of size 0 in subroutine glide_allocarr in glide_types.F90.

The simplification requires eliminating the ability to perform math on multiple model variables in the definition of the 'data' that defines output variables in glide_vars.def.  The only output variable that currently used that type of definition is 'slc'.  After consulting with other CISM developers, I have removed the 'slc' variable from glide_vars.def.  If it is desired in the future, it should be added as a model variable in glide_types.F90 and have its data specified as a single model array, rather than the difference of two arrays.

I have tested this with a handful of common options (e.g. glide, glam, glissade dycores) and encountered no errors.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@998 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 9a24bbf46b05b24eb961d49b90ae4e183c9f01fd