Modernization of the dome.py script.
Commanline options are parsed with the optParse module.  Script can be executed with --help for details of allowable options.
If a parallel run is requested the script tries to determine which of 'openmpirun', 'mpirun', 'aprun' to use.
Also deleted some unnecessary stuff - like the commands to copy to the nonexistent 'scratch' dir that always throws an error.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1006 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 0528401e19dda66a6cfa9b7ab748309c23393f4d