Added a new subroutine, test_trilinos, for testing the new trilinos solver in glissade_velo_higher.F90.
To run the test:  Set trilinos_test = .true. near the top of subroutine
glissade_velo_higher_solver.  Build the code in parallel with trilinos,
then launch it with either 1 or 2 processors.  With 1 processor, the
test matrix is 2x2.  With 2 processors, the test matrix is 4x4.
Currently the solve fails in solvevelocitytgs at the FillComplete step.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1030 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: cdc9bc4a9727bd1e3da48daa3d8f9ad3ddcf72a7