Some changes related to glissade solve with Trilinos:
Added an additional matrix test problem and verified that the solution is
correct (1) in parallel and (2) for consecutive solves with different numbers of unknowns.

Cleaned up some allocate/deallocate statements.

Verified that for the diagnostic dome problem, the Trilinos solver converges
to the same solution as the native PCG solver (to 12 decimal places) in the same
number of nonlinear iterations (51).  I did this on a Mac for 1, 2 and 4 processors.

However, the Trilinos solver is very slow, and each linear solve yields a message
like the following:
Precon Linear Solve Failed: 4.75e-05 drop in 200 its (avg: 193 its/slv, 5.88% success)

So something is not quite right.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1032 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 64b0df668cd68e3a61d23df25ae2b16744c1daeb