Modified the logic for the glissade-Trilinos solver so that all columns that can potentially be nonzero are communicated to Trilinos for each active row.  Previously I was leaving out some values that were zero initially but then changed to nonzero.
The solver now successfully converges for the ISMIP-HOM A and C tests
for all the problem sizes and grid dimensions I checked, including 5-km.
Answers are the same as for the slap and native PCG solvers, to within roundoff.

Based on discussions with Andy, I may be passing too many columns to Trilinos
for boundary nodes.  This remains to be sorted out.  But the code is working for now.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1043 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 1918a24e5961c6f0d4dbba11504e06d759ee84af