More mods to glissade_velo_higher to support the Trilinos solve.
At the beginning of each timestep, I fill a logical array that specifies
whether each neighboring node is active or inactive for each active node.
This array is used during assembly to determine which (row, column) matrix
values are potentially nonzero and should be passed to Trilinos.

This is a better implementation than revision 1043, which allowed values
from inactive columns to be passed to Trilinos.

The Trilinos solver is still running slowly, but the answers are the same
(within roundoff) as for the SLAP and native PCG solvers for ismip-hom A
and dome test cases.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1056 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 4a47442a0ce4933f242c9db16aa6452b729db515