Created new subroutine to check for advective and diffusive CFL violations on each time step prior to advection for higher-order dycores.
The check occurs in glissade_tstep() prior to evolution actually occurring, and accounts for subcycling (if enabled).

Currently an advective CFL violation is a fatal error, but a diffusive CFL violation is just a warning (written to the log).  This is because the DCFL calculation assumes all flux in the downslope direction is diffusive.  This will be overly restrictive in stream-y cases so we don't want this to be fatal.  However, the diffusivity calculated here using the glissade dycore with no sliding for the dome test case matches within 5% the SIA-diffusivity calculated by glide for the same test case.  Thus, this DCFL check in its current formulation will primarily be useful for assuring stable evolution for cases with little sliding.  Eventually, we will try to make a 'smarter' DCFL calculation that is practical for cases with rapid sliding.

I have also created scalar output variables called adv_cfl_dt & diff_cfl_dt that can be used to output these values to the output .nc file.  If included in the output file, you can do something like:
ncdump -v adv_cfl_dt,diff_cfl_dt dome.out.nc
to see their values for each timestep (or use ncview).  This is useful to see how close to the values your chosen dt is, since the warnings/errors described above only occur with a CFL violation.  This could also reveal if something about your geometry changed that requires shorter time steps.  make clean is necessary to have access to the output variables.

Note: the value of the maximum allowable time step in the log messages generated by this commit may be inaccurate for parallel runs - this will be fixed in followup commits. (but the logic determining if an error occurs is still valid)

Note: Eventually we may use these calculations to make the dt adaptive, but that is not a near-term goal.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1103 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 77f735b027245388916d480b8b3fec8f1760aec9