Added an interface for parallel_reduce_maxloc, analogous to parallel_reduce_minloc.
Added a subroutine called parallel_localindex.  This is the inverse of parallel_global_index.
Given the global i and j indices, it determines the processor that contains the global point
and the local i and j indices of that point. If the global indices are invalid, the code
aborts with an error message.

Modified glide_diagnostics.F90 to use the new parallel_localindex subroutine.

Verified that these new subroutines have the expected behavior in parallel (on 1, 2 and 4 procs)
and in serial.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1130 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 8b78ebe703547248e1a69aa3b21762ff7b0382bc