I modified glissade_velo.F90 and glissade_velo_higher.F90 so that calcbeta is called from inside glissade_velo_higher_solve.  This requires setting up some additional pointers in glissade_velo_higher_solve and converting to and from the dimensionless units expected by calcbeta.  I verified that results are the same (within roundoff) for several values of which_ho_babc.
Note: Because of dimensional conversions (dividing and then multiplying by a constant),
answers for the glissade dycore have changed at the roundoff level.
In a followup commit, I will modify calcbeta so that it expects input in SI units
instead of dimensionless, so these conversions will not be needed.

In glissade_velo_higher.F90, I modified subroutine dirichlet_boundary_conditions
to handle inhomogeneous BC.  This requires subtracting some terms from the RHS
vector that previously were ignored.  The unmodified load vector is now called
loadu/loadv, and the modified vector (with Dirichlet terms subtracted) is bu/bv.

I added a field model%geometry%stagmask to replace geom_mask_stag, which was not
part of the model derived type.

I began testing the stream test case with the glissade velocity solver:

* I verified that the code converges to the correct velocity field when supplied
with a prescribed beta field and an inhomogeneous Dirichlet BC (velocity
prescribed at the east and west edges of the domain).

* For the doubly periodic stream test with prescribed yield stress mintauf,
I got the code to converge, but the maximum velocity is only ~5 m/yr, compared
to ~900 m/yr for the Raymond analytic solution.  Still have to figure this out.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1195 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 1ecf9d29f09870b1b2c83599ee68a82a52b285a2