I modified calcbeta so that the input/output arguments are assumed to be dimensional: e.g., Pa yr/m for beta, Pa for mintauf, m for bwat, and m/yr for uvel and vvel. This means that dimensional conversions before and after the call to calcbeta are no longer required for the glissade solver, but are now required for the glam solver.
NOTE ON ANSWER CHANGES: Since beta is now divided and multiplied by a constant within glam,
results could change at the roundoff level.  My dome results did not change on my Mac,
but results could change on other platforms or tests.

I verified that glissade results are the same (within roundoff) for the following tests:
    ismip-hom A, which_ho_babc = 0, 4, 6
    ismip-hom C, which_ho_babc = 5, 8
Likewise, I checked that glam results are the same (within roundoff) for these tests:
    ismip-hom A, which_ho_babc = 0, 4
    ismip-hom C, which_ho_babc = 5
    stream, kinematic BC, which_ho_babc = 2
    stream, periodic BC, which_ho_babc = 2
For reasons I'm unsure of, glam previously (and still) does not converge for which_ho_babc = 6 or 8.

The glissade solver still converges to a strange result for the periodic stream case.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1196 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 1f688a1019222df60507adb12e37c1af05d08167