Minor update to the staggered_parallel_halo_extrapolate_integer_2d and staggered_parallel_halo_extrapolate_real8_2d routines to improve bit-reproducibility on multiple processors.
Modified code to now copy entire east and west columns rather than
just the owned cells in those columns.  This avoids having the halos have
potentially wrong information (i.e., a few cells in the corner don't get extrapolated into)
I think a halo update after this subroutine would also solve the problem,
but this avoids an MPI-communication (assuming the halos are up-to-date
prior to this subroutine).

This reduces the bit-reproducibility problem with the stream test case
(pushes the difference out from the 1e-3 to the 1e-5 level) but does not completely
eliminate it - there appears to still be a problem somewhere else that I have been unable to locate.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1212 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 89a81f1a98b1e811c89527ad979a75d2550390e1