ISMIP-HOM test case
Updates to run and plot scripts to get experiment F running and plotting.  The problem is still not setup correctly, because velocities are about half what they should be, but the scripts function successfully, and the shape of the velocity profile looks pretty good, so I think it is close.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1224 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 141547cd6c301031934898d594d6718004e304bb