ISMIP-HOM Test Case
Moved the code that generates the standard format ascii output files from the netcdf output file from the run script to the plot script.  This is handy for doing runs on remote machines that do not have the necessary modules to use the plot script - now you can just copy the .out.nc file to your local machine and use the plot script without a problem.  (This also removes numpy as a requirement from the run script, which makes it more portable.)

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1235 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: a35494986e20810d1457cd701f52ed03f650aa44