Adding wvel_ho
Adding a new diagnostic variable called wvel_ho which is the vertical velocity relative to the ice sheet base for higher-order dycores.  The use of a different variable name 'wvel_ho' is to distinguish it from the SIA variable wvel which is the vertical velocity on a fixed grid (the difference only matters if lsrf is changing due to isostasy or floatation but could be potentially confusing).  Also, wvel_ho can currently only be computed with the VERTINT_STANDARD method - the VERTINT_KINEMATIC_BC method is not currently supported.  Both of these restrictions could be relaxed as needed (and then the two variables could be merged into one), but would require calculating time derivatives, which is avoided at present.  Currently, the only use of wvel_ho is to compare to standard solutions in the ISMIP-HOM test cases (figures in the supplementary material).

wvel_ho is allocated only for higher order dycores.  'wvel_ho' has been added as an output variable in glide_vars.def - make clean required to use it.

Also modified the ismip-hom F test to make use of the new variable.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1258 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 3ed882e5e0af884e532510d79f8b748b70c0f460