Change the conditionals that get the trilinos compilers to only do this if neither CMAKE_Fortran_COMPILER nor ENV{FC} are defined (and similar for CC and CXX). This is needed for the CESM build, which sets the compilers via the FC, CC and CXX environment variables, rather than via CMAKE_Fortran_COMPILER, etc. (and we don't want to use the trilinos compiler if it was built using an older compiler version).
There is a slight chance that this could break the build on some platforms. This
could happen if (1) you are NOT explicitly specifying the CMAKE_*_COMPILERs in
your call to cmake, and (2) you have defined the CC, CXX and/or FC environment
variables (perhaps inadvertently), and (3) you have been relying on cmake to get
the proper compiler for you based on the trilinos build. Now it would use CC /
CXX or FC instead of grabbing the compiler from trilinos. If this is the case,
then I recommend changing your cmake invocation to explicitly specify the
compilers you want to use rather than relying on trilinos to give you this
information.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1316 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: ed7ebfb6533b897e09cb681f16800bc42fddbc51