Changes related to testing and debugging of the first-order upwind transport scheme
In glissade_transport.F90, I fixed a bug that had the effect of transporting
ice with twice the actual velocity.  This changes the answers, of course, but
to my knowledge no one has been	using the upwind scheme (evolution = 4) in any
standard tests.  Answers using incremental remapping (evolution = 3) are the same.

I separated an old glissade.F90 subroutine, glissade_test_parallel, into two parts:
glissade_test halo and glissade_test_transport.
I then modified and expanded glissade_test_transport so that it can be used
to transport an initial cylinder (or other shape) of ice exactly once around the
domain back to the starting point.  Comparing the transported to the initial
geometry is an easy way	to evaluate transport schemes. I verified that
IR and 1st order upwind	behave as expected: upwind is much more	diffusive
when there is a	 sharp ice edge.  Differences between the two transport schemes
are smaller when the initial geometry is smoother, as for the dome and
Halfar problems.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1320 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: b90e48fbedb4249e84a9033b1b309273dbb008cd