Changes	related	to glissade's native PCG solver.
Thanks to Pat Worley for finding that the new Chronopoulos-Gear	solver is not
behaving as desired.  It sometimes requires more linear iterations to converge
than does the standard PCG solver.  This should	not be the case, so there is
likely a bug.  For now I've backed off to the standard PCG solver as the default
(set in	glissade_velo_higher.F90).

NOTE: As a result of reverting to the standard PCG solver, answers will change
at the roundoff level for glissade (dycore = 2) with the native PCG solver
(which_ho_sparse = 3).

Also in	glissade_velo_higher.F90, I fixed a spelling mistake in a call to
timer glissade_load_vector_gravity.

In cism_sparse_pcg.F90, I added	a subroutine solve_test_matrix_standard	so I
could compare the standard and Chron-Gear solves for simple matrices.
I've found no differences so far.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1321 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 5b679c95f0aac5b3dac64ff169cb17328340b5ef