BRANCH COMMIT - forcing
First cut at functioning time-dependent forcing.

To use, add a section to the config file like:
[CF forcing]
name = dome.forcing.nc

Multiple forcing file sections should work in principle (not tested).

The current functionality reads all forcing fields present in each forcing file on every time step (i.e., a read from file occurs even if it does not result in a changed forcing value).  It takes the latest time level from the forcing file(s) less than or equal to the current time (stepwise constant forcing), but there is no epsilon applied for roundoff error so that does not always work quite right yet.

The implementation makes a new variable inside the glide_funits type called "frc_first" that is a linked list of forcing files.  It has the same datatype as input files (glimmer_nc_input) since the two require very similar behavior.  I think this datatype is flexible enough for forcing needs, but if not, a separate forcing type could be created.  The downside to that would be the need to re-create/modify all of the autogenerated read routines (not fun).

The actual forcing is read by a subroutine called NAME_read_forcing in the *_io.F90 autogenerated files (i.e., in ncdf_template.F90.in).  It's unclear to me if this routine makes any sense for *_io modules besides glide_io.F90 (e.g., for glint_io.F90).  However, the glide_read_forcing() is the only version that is actually called anywhere (which occurs in cism_run_dycore).  This might be an issue to followup on before merging this into the trunk.

Handling of the config options associated with the [CF forcing] section happens in glimmer_nc_params.F90 (routine handle_forcing and associated code).  If we add the ability to do, e.g., linear interpolation in time, or to choose what happens when the forcing file runs out (repeat file from beginning or hold last value steady in time), those choices should be handled here.  Note that there are some options associated with the glimmer_nc_input data type that don't make sense to read from the config file for forcing (time=1).  There currently is no handling of those situations.

The forcing files are opened in glimmer_ncio.F90 when input and output files are opened.  The forcing files are currently not closed which will need to be dealt with in a followup commit.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/time_dependent_forcing@1405 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 1a3d3c5a2de5e798a0803e9d9aa164e2a9751504