In this version, references to anything in example-drivers have been commented out.  This means the old drivers simple_glide, simple_bisicles, glint_example and eis_glide will not build.  Their functionality has been duplicated in cism_driver.
I tested this on a mac, and saw no problems.  I'll test on other systems as well.
If there are problems, these changes are easy to back out.  If no problems are seen,
the example-drivers directory can be deleted.

Note that the cmake variables CISM_BUILD_* haven't been renamed yet.  These are
now used to build cism_driver to match the functionality of the corresponding old
driver.  For example, setting CISM_BUILD_SIMPLE_BISICLES will produce a version of
cism_driver that duplicates simple_bisicles functionality.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1425 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 7fdb2506345330c125fa500a697f6e094bebaa5c