In glide_transport.F90 I replaced the old vertical remapping routine, whose computational cost scales as nlyr^2 (where nlyr is the number of vertical ice layers), with a new routine that gives the same answer but whose cost scales as nlyr.
The reduction in cost is trivial with 10 vertical layers but would be greater if
the code is run with a larger number of layers.

I verified that the answers are BFB for the dome evolution test case.
If anyone finds that answers have changed for another problem, please let me know.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1508 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: bd3dbe9c3cc9c9de0c9cec37c54d37f62b46607f