Fixing scripts for Ross Test Case
I redid the grid conventions for this test case: 1) the extra 2-cell padding that was a holdover from the early days of incremental remapping has been removed; 2) I chose to assume that the supplied data for the test case is on the velocity grid (topg and thk are interpolated from the supplied grid back onto the unstaggered grid).

I cleaned up the runRoss.py script after these changes to eliminate options and logic that are no longer needed.  I cleaned up plotRoss.py to use these new conventions, and also fixed a bug that was preventing the output from being scaled when using the Scientific.IO.NetCDF module.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1532 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 3c54858103b70822acff39f5811d90a405a3db38