Additions to the glissade velocity solver to enable a 2D shallow-shelf solve.
I also cleaned up and rearranged some glissade modules and subroutines,
and created several new glissade modules containing code that was previously in
glissade_velo_higher.F90.

A fresh configure and build will be needed to use the new modules.

Answers are unchanged for the standard glissade tests using the 3D higher-order
(Blatter-Pattyn) solver.  Answers are changed for the SSA solver (which_ho_approx = 1),
but no standard tests use this solver.

The glissade velocity solver can now solve a well-posed SSA problem efficiently
in 2 dimensions for the various shelf test cases.  Details are below.

****************************************

The previous SSA implementation in glissade_velo_higher was flawed.
When assembling the matrix, I ignored vertical shear terms in the effective viscosity
and in the element stiffness matrices.  The first assumption was correct, but
the second led to an ill-posed problem.

If I retain the vertical shear terms in the stiffness matrix, then the
problem is well-posed and I get a good solution. However, it is still a 3D problem that is
no faster or easier to solve than the full HO problem.  So I decided to modify the
SSA solver to do a 2D solve at a single horizontal level.

If starting from scratch, I would assemble a global matrix for the 2D problem.
But this would require a lot of extra coding.  Instead, I now assemble the
global matrix for a 3D problem, and then make the assumption that there is
no vertical shear in uvel or vvel.  This allows matrix terms to be gathered
and summed over columns such that the resulting problem is 2D (i.e., we solve
for vertically averaged uvel and vvel at a single level).

I can then use the existing solvers (with straightforward modifications to exclude
the vertical dimension) to solve the 2D problem. Note, however, that SIA-based preconditioning
cannot be applied to a 2D shelf problem. If a user tries to do this, the code
will abort with a fatal error message at initialization.

The 2D shallow-ice matrix has 1/N as many rows at the 3D HO matrix,
and the number of columns per row is reduced by a factor of 3 because there
is no coupling to levels above and below.  As a result, the SSA solve is now
much faster than the HO solve for a given shelf problem.

I checked results for the circular shelf, confined shelf, and Ross shelf problems.
I verified that when using the SSA (which_ho_approx = 1), all the sparse solvers
(SLAP, PCG and Trilinos) give the same answers within roundoff.

Also, I verified for these shelfy problems that the SSA results are similar to the HO results.
Where there are differences, these can be explained by the boundary conditions.  For instance,
the circular shelf problem nails down the shelf in the center by prescribing a large
value of beta, which means the SSA is not a good approximation at the shelf center.

I added several new solver subroutines to handle the new 2D data structures.
(These are very similar to the existing 3D solver subroutines.)
This made glissade_velo_higher.F90 even longer than it was before, so I decided to
move some subroutines to new modules.  Now there are modules called
glissade_velo_higher_slap.F90 and glissade_velo_higher trilinos.F90; these subroutines
process arrays and call solvers for the SLAP and Trilinos solves, respectively.
Also, I moved libglimmer-solve/cism_sparse_pcg.F90 to libglissade/glissade_velo_higher_pcg.F90,
since this module is glissade-specific and is unrelated to the other modules in
libglimmer-solve.

In addition, I created a module called glissade_masks.F90 with some mask subroutines,
including the new code to compute the grounded fraction in each grid cell.

I also removed several obsolete subroutines from glissade_velo_higher.F90.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1571 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: e5daeec8ce792d2427e20f918e55e0855211618d