Hopefully the last mass-edit of these scripts.  I used the updated mac-gnu-cmake script as a template for the other scripts, replacing the first 10 lines of the cmake calls (a bunch of switch settings) with a new standard set.  I also updated the multiple lines of comments that describe these settings.
I tested on hopper with hopper-gnu-cmake, but haven't tested the other 21 scripts.  I tried
to minimize the possiblity of breaking builds, but if you encounter a broken build, let me know.

Also, if you created a build script that doesn't end with -cmake, it hasn't been updated, so
it is probably broken.  I can help fix those, too.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1622 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 6a6841643a31b9cea236d3de083e2fc7f89ac46e