For the shelf test case I reverted the Trilinos linear solve threshold from 1.e-08 to 1.e-11, to be consistent with the other higher-order test cases.
This threshold might be smaller than necessary.  If so, we could raise it
at some point (consistently for all the test cases).

Answers will change at the roundoff level for the circular-shelf and
confined-shelf test cases using Trilinos, but I don't think any standard
test cases use Trilinos for the shelf problems.

I fixed a comment in the shelf README file, fixed a typo in the Halfar README
file, and added a comment in runRoss.py.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1626 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 5b80ae445486092efd69fa4b64fa17f517595a97