Created a new parameter, efvs_constant, in the glide_paramets derived type. This is the effective viscosity used for the case which_ho_efvs = HO_EFVS_CONSTANT = 0.
The default value is 2336041.42829d0 as prescribed for ISMIP-HOM Test F.

The default value can be overridden by setting efvs_constant in the [paramets]
section of the configure script.

Code behavior is the same as before (answers are BFB), but now the user
can override the default value without editing source code.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1637 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: c66e312d476b9c3ec2c96311b17fd73cba3bd320