Changed efvs_constant from 2336041.42829d0 Pa Yr to 2336041.d0 Pa yr.
This value is equal to 0.5 * A^(-1), where A = 2.140373 Pa^(-1) yr^(1) is the value used in ISMIP-HOM Test F.
Since there are 7 sig figs in this prescribed value of A, it seems appropriate to use 7 sig figs in efvs_constant
rather than 12.

Answers change at around the 7th decimal place for tests with which_ho_efvs = 0 (constant efvs).
I don't know of any standard tests with constant efvs.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1638 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: f9be4bc297e79ecc48bffa4fb1f1ace1e57cb9dd