Fixing fatal error (that occurs on some builds) when glint tries calling glide_nc_fillall() to setup glint output files.
On my machine I get a fatal error, but Bill's does not (the difference *may* be related to differences in netcdf version).  Nonetheless, we came up with a fix which is to add the glide dimension variables that are getting added in that routine to the glint_mbal_vars.def file.  These dimension variables are not used at all by glint, but this eliminates the error.  A more appropriate fix might be to create a glint_nc_fillall() that would not try to write these variables to the output file.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1641 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 231294f8bc07a4b75a403958b782be61ea56cbcb