Edits connected with glint and the glint_example test case.
I verified that the glint_example test case is working with cism_driver for
single-processor Glide runs, with forcing in both PDD and SMB modes, on a
20-km Greenland grid.

For the PDD case I confirmed that answers are the same within roundoff
compared to runs from a March checkout (r1114).  This was around the last
time glint_example was tested.

For the SMB case the answers are different from r1114. I traced the differences to
the new treatment of the SMB over bare land, as introduced by Bill Sacks's
commit on May 2 (r1289).  These differences are to be expected.  With the
new treatment, glint_example in SMB mode does not permit glacier inception
over bare land (except indirectly--ice from a positive SMB can begin to
accumulate once the ice expands dynamically into a formerly ice-free region).
For testing purposes I think this is OK.

I started testing glint_example with glissade and have run into some issues:

(1) The code was finishing prematurely after the first call to glint.
    I fixed this by editing subroutine gcm_to_glint_finished in gcm_to_cism_glint.F90.
    Previously this subroutine could return 'finished' = T even if the
    finish conditions were not met, if 'finished' were T on input (which
    was happening for some reason). Now it returns the correct value of
    'finished' regardless of the input value.

(2) There is a transport conservation error on the first time step, even
    when glissade is run on one processor.

(3) There is a seg fault when glissade is run with glint_example on more
    than one processor.

I'll try to fix (2) and (3) with future commits.

In several other modules I cleaned up comments and made other minor,
non-answer-changing edits.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1647 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 429283a0c84309a65723bf3cfc6d17e8e8074707