BRANCH COMMIT - friction laws
Adding effecpress (on the scalar grid) as another variable in the basal_physics type.
This variable can be loaded from an input file.
Near the beginning of glissade_diagnostic_solve() effecpress is staggered to calculate the value of effecpress_stag that is actually used by the friction laws.
I have added effecpress to the restart list if either of the two new friction laws are specified for which_bo_babc in the config file - otherwise beta is added to the restart list instead.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/branches/coulomb_friction_law@1668 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 98aba93b34322cace2f4dacda5c15ed9936573bb