Various changes related to the glint_example driver, as well as some general cleanup.
Answers are unchanged for standard test problems (as verified by running LIVV for Mac),
but will differ for glint when run with glissade.

The glint_example driver (launched by cism_driver followed by two filename arguments)
is now working for both glide and glissade.  It works for glissade with SMB forcing
with both serial and parallel builds.

Details are below.

**************

Previously, glint_example runs with glissade using SMB forcing gave different answers in serial and parallel.
Both serial and parallel runs had an energy conservation error.  And the model died with a seg fault
when running glint_example with glissade using more than 1 processor.

I made these fixes:

* I added a number of changes previously made by Bill Sacks (but not committed) in glint_downscale.F90
and glint_interp.F90, to correctly downscale global fields to the local grid in the bare-land
elevation class (elevation index 0).  These include a new subroutine called copy_to_local.

* I found and fixed a bug in the gather and scatter subroutines of parallel_slap.F90.
These subroutines were assuming equal sizes for the global and local arrays, which works only
for the case of no halo rows (as in glide but not glam and glissade).  Now the serial
subroutines work correctly when nhalo > 0.

* In subroutine interp_to_local of glint_interp.F90, before scattering fields from the global
to local arrays, I initialized the local arrays to zero to make sure there are sensible values
in the halo cells (which are not filled by the scatter).  I considered removing the parallel_halo
calls after the scatter, but kept them for now just in case some subroutine assumes these
fields are correct in the halo.

* In glissade_transport.F90, I found and fixed a bug in the energy conservation diagnostic,
which was reporting non-conservation in some cases where energy was in fact conserved.
(These were cases where the surface mass balance was negative enough to melt all the
ice in a given column.) I revised the conservation checks to include separate checks
for three procedures: horizontal remapping, vertical remapping, and adding the surface
and basal mass balance. I also introduced a computation for unused melt potential in cells
where the negative mass balance is more than sufficient to melt the entire ice column.
Later it might be useful to pass this melt potential back to a climate model to ensure
exact conservation with two-way coupling.

Some other changes:

* I introduced a new global_sum interface in the parallel modules.

* I cleaned up some diagnostic prints that were occurring on all processes instead
of just main_task, or were taking place on every single timestep.

* I slightly changed the interface for glissade_transport_driver.

* The code now aborts if a user tries to run glint with the PDD scheme on more than
one processor.  It might be possible to get this to work, but it's not a high priority.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1713 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 53b49ab007c364d9b490b434bebb540003045e88