Remove some unnecessary pointer assignments in order to address problems with the intel compiler, which for some reason didn't like it when we tried to deallocate the arrays that had been involved in these pointer assignments.
Previously, the pointer assignments had looked like this:

  btraction => fptr%velocity%btraction(:,:,:)

  ! a bunch of code here

  fptr%velocity%btraction => btraction(:,:,:)

This second set of pointer assignments is unnecessary, because
btraction and fptr%velocity%btraction already point to the same
memory. So I removed this second set of pointer assignments. This
allows the deallocations to work with the intel compiler, so I have
restored these deallocations that were previously not being done.

I tested these changes with LIVV on titan with the latest code as of
Sept. 3. These files have only changed minimally on the trunk since
then, so I have not retested with the latest trunk.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1715 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 101a9c877e4a299544a317f6f406865265f3747e