Minor bug fix for call to calcbeta in glissade_velo_higher.F90.
This subroutine was being called with input arguments uvel/vvel(nz,:,:),
which had not been filled for the SSA and L1L2 options (whichapprox = 1 and 3).

Now the input arguments are correct for all the whichapprox options, and the stream
test case is working for SSA and L1L2.  The SSA and L1L2 results actually look slightly
better than the BP results.

I confirmed with Mac LIVV that standard test results (using Glissade BP) are BFB.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1760 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 99a4cd9282c0fbc6b181213c9e5ca8fbb8a668d5